Rewrite Wayland support.
• Apply 'os flags' automatically (without any configuration).
• Do not try applying native Wayland automatically, it seems to be
  currently not possible anyway.
• Enable native Wayland when `--ozone-platform=wayland`.
• Automatically enable PipeWireCapturer for both XWayland and
  native wayland.
• Prepare desktopCapturer to support PipeWireCapturer.

Resolves: Wayland screen capture support #100.

+++ sources/code/main/windows/main.ts
@@ -296,40 +296,54 @@ export default function createMainWindow(startHidden: boolean, l10nStrings: l10n
                 // Fail when client has denied the permission to the capturer.
                 if(!configData.get().permissions["display-capture"])
                     return new Error("Permission denied.")
-                lock = true;
+                lock = !app.commandLine.getSwitchValue("enable-features")
+                    .includes("WebRTCPipeWireCapturer") ||
+                    process.env["XDG_SESSION_TYPE"] !== "wayland";
                 const sources = desktopCapturer.getSources({
-                    types: ["screen", "window"],
-                    fetchWindowIcons: true
+                    types: lock ? ["screen", "window"] : ["screen"],
+                    fetchWindowIcons: lock
                 });
-                const view = new BrowserView({
-                    webPreferences: {
-                        preload: resolve(app.getAppPath(), "app/code/renderer/preload/capturer.js")
-                    }
-                });
-                ipcMain.handleOnce("getDesktopCapturerSources", (event) => {
-                    if(event.sender === view.webContents)
-                        return sources;
-                    else
-                        return null;
-                });
-                const autoResize = () => setImmediate(() => view.setBounds({
-                    ...win.getBounds(),
-                    x:0,
-                    y:0,
-                }));
-                ipcMain.once("closeCapturerView", (_event,data:unknown) => {
-                    win.removeBrowserView(view);
-                    view.webContents.delete();
-                    win.removeListener("resize", autoResize);
-                    resolvePromise(data);
-                    lock = false;
-                })
-                win.setBrowserView(view);
-                void view.webContents.loadFile(resolve(app.getAppPath(), "sources/assets/web/html/capturer.html"));
-                view.webContents.once("did-finish-load", () => {
-                    autoResize();
-                    win.on("resize", autoResize);
-                })
+                if(lock) {
+                    const view = new BrowserView({
+                        webPreferences: {
+                            preload: resolve(app.getAppPath(), "app/code/renderer/preload/capturer.js")
+                        }
+                    });
+                    ipcMain.handleOnce("getDesktopCapturerSources", (event) => {
+                        if(event.sender === view.webContents)
+                            return sources;
+                        else
+                            return null;
+                    });
+                    const autoResize = () => setImmediate(() => view.setBounds({
+                        ...win.getBounds(),
+                        x:0,
+                        y:0,
+                    }));
+                    ipcMain.once("closeCapturerView", (_event,data:unknown) => {
+                        win.removeBrowserView(view);
+                        view.webContents.delete();
+                        win.removeListener("resize", autoResize);
+                        resolvePromise(data);
+                        lock = false;
+                    })
+                    win.setBrowserView(view);
+                    void view.webContents.loadFile(resolve(app.getAppPath(), "sources/assets/web/html/capturer.html"));
+                    view.webContents.once("did-finish-load", () => {
+                        autoResize();
+                        win.on("resize", autoResize);
+                    })
+                } else {
+                    sources.then(sources => resolvePromise({
+                        audio: false,
+                        video: {
+                            mandatory: {
+                                chromeMediaSource: 'desktop',
+                                chromeMediaSourceId: sources[0]?.id
+                            }
+                        }
+                    })).catch(error => console.error(error));
+                }
                 return;
             });
         });
